+++ ksp/src/jvmMain/kotlin/eu/vendeli/ksp/dto/CommonAnnotationData.kt
@@ -26,7 +26,7 @@ sealed class CommonAnnotationValue {
             is String -> "CommonMatcher.String(value = \"$value\"$parametersString)"
             is Regex ->
                 "CommonMatcher.Regex(value = Regex(\"$value\"${
-                    value.options.takeIf { it.isNotEmpty() }?.joinToString(prefix = " ,") { "RegexOption.$it" } ?: ""
+                    value.options.takeIf { it.isNotEmpty() }?.joinToString(prefix = " , setOf(", postfix = ")") { "RegexOption.$it" } ?: ""
                 })$parametersString)"
         }
     }