+++ telegram-bot/src/commonMain/kotlin/eu/vendeli/tgbot/types/configuration/BotConfiguration.kt
@@ -5,7 +5,9 @@ import eu.vendeli.tgbot.implementations.ClassManagerImpl
 import eu.vendeli.tgbot.implementations.InputListenerMapImpl
 import eu.vendeli.tgbot.interfaces.ctx.ClassManager
 import eu.vendeli.tgbot.interfaces.ctx.InputListener
+import eu.vendeli.tgbot.interfaces.helper.Middleware
 import eu.vendeli.tgbot.types.component.ExceptionHandlingStrategy
+import eu.vendeli.tgbot.utils.common.ProcessingCtxKey
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 
@@ -20,6 +22,8 @@ import kotlinx.serialization.Transient
  * @property inputAutoRemoval A flag that regulates the auto-deletion of the input point during processing.
  * @property exceptionHandlingStrategy Exception handling strategy. See [ExceptionHandlingStrategy].
  * @property throwExOnActionsFailure Throw exception when the action (any bot request) ends with failure.
+ * @property processingCtxTargets List of targets for which the context will be automatically tracked.
+ * @property middlewares List of middlewares that will be applied to all requests.
  */
 @Serializable
 @ConfigurationDSL
@@ -35,6 +39,9 @@ data class BotConfiguration(
     @Transient
     var exceptionHandlingStrategy: ExceptionHandlingStrategy = ExceptionHandlingStrategy.CollectToChannel,
     var throwExOnActionsFailure: Boolean = false,
+    val processingCtxTargets: MutableSet<ProcessingCtxKey> = mutableSetOf(ProcessingCtxKey.REGEX_MATCH),
+    @Transient
+    val middlewares: MutableList<Middleware> = mutableListOf(),
     @Transient
     internal var rateLimiter: RateLimiterConfiguration = RateLimiterConfiguration(),
     internal var httpClient: HttpConfiguration = HttpConfiguration(),
@@ -77,21 +84,23 @@ data class BotConfiguration(
         commandParsing.block()
     }
 
-    internal fun apply(new: BotConfiguration): BotConfiguration {
+    internal fun rewriteWith(new: BotConfiguration): BotConfiguration {
+        identifier = new.identifier
         apiHost = new.apiHost
+        isTestEnv = new.isTestEnv
         inputListener = new.inputListener
         classManager = new.classManager
+        inputAutoRemoval = new.inputAutoRemoval
+        exceptionHandlingStrategy = new.exceptionHandlingStrategy
+        throwExOnActionsFailure = new.throwExOnActionsFailure
+        processingCtxTargets.addAll(new.processingCtxTargets)
+        middlewares.addAll(new.middlewares)
         rateLimiter = new.rateLimiter
         httpClient = new.httpClient
         logging = new.logging
         rateLimiter = new.rateLimiter
         updatesListener = new.updatesListener
         commandParsing = new.commandParsing
-        identifier = new.identifier
-        isTestEnv = new.isTestEnv
-        inputAutoRemoval = new.inputAutoRemoval
-        exceptionHandlingStrategy = new.exceptionHandlingStrategy
-        throwExOnActionsFailure = new.throwExOnActionsFailure
 
         return this
     }